Change method of data filtering in methylation binning
Change data reading method to read in the data when setting methylation
attributes. Rather than Reading all data in at once and filtering down
the dataframe for each gene, use the first line as a string to get the
column headers within each gene and pass to usecols to read in only the
necessary columns for each gene instance. Additionally, add features to
retain starnd information for both gene and methylation sites, and change
"upstream" and "downstream" to "pre" and "post" to reflect the lack of
incorporation of strand information in the calculation of the bins.